% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{calculate_CNV}
\alias{calculate_CNV}
\title{Determine CNV breakpoints by BICseq2 and output single cell copy ratio profile}
\usage{
calculate_CNV(
  norm_count,
  baseline,
  BICseq = "default",
  lambda = 5,
  min_interval = 2,
  smooth = F,
  downsample = 0,
  output_dir = "./",
  output_name = "CNV_result.rds"
)
}
\arguments{
\item{norm_count}{a normalized read count matrix. Obtained from previous step
\code{normalize()}.}

\item{baseline}{read count baseline of normal cells. Obtained from previous step
\code{normalize()}.}

\item{BICseq}{path to BICseq executable file.}

\item{lambda}{parameter of BICseq segmentation algorithm that controls the overall
resolution of segmentation result. In general, more breakpoints will be reported
when smaller \code{lambda} is used.}

\item{min_interval}{minimum bin number between breakpoints.}

\item{smooth}{logical. Whether to aggregate similar cells to smooth final CNV
results?}

\item{downsample}{an integer. If larger than 0, rows of input matrix will be
randomly down-sampled to this number to accelerate segmentation. 1000 is
recommended if the sample contains too many cells.}

\item{output_dir}{output directory.
The generated temporary files will also be output to this directory.}

\item{output_name}{output file name, should be ".rds"}
}
\value{

}
\description{
Determine CNV breakpoints by BICseq2 and output single cell copy ratio profile
}
\examples{

}
