% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{calculate_CNV_arm_level}
\alias{calculate_CNV_arm_level}
\title{Arm-level CNV}
\usage{
calculate_CNV_arm_level(
  norm_count,
  baseline,
  genome = "hg38",
  min_interval = 5,
  smooth = F,
  output_dir = "./",
  output_name = "CNV_result.rds"
)
}
\arguments{
\item{norm_count}{a normalized read count matrix. Obtained from previous step
\code{normalize()}.}

\item{baseline}{read count baseline of normal cells. Obtained from previous step
\code{normalize()}.}

\item{genome}{reference genome, "hg19", "hg38" or "mm10".}

\item{min_interval}{minimum bin number between breakpoints.}

\item{smooth}{logical. Whether to aggregate similar cells to smooth final CNV
results?}

\item{output_dir}{output directory}

\item{output_name}{output file name, should be ".rds"}
}
\value{

}
\description{
Arm-level CNV
}
