% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{del_chr}
\alias{del_chr}
\title{Delete specific chromosomes}
\usage{
del_chr(count, chr_list = c("chrX", "chrY"))
}
\arguments{
\item{count}{a cell-by-bin matrix with formatted column names}

\item{chr_list}{a list of chromosomes to remove}
}
\value{
a cell-by-bin matrix
}
\description{
Delete specific chromosomes
}
