% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_copy_state.R
\name{estimate_cnv_state_cluster}
\alias{estimate_cnv_state_cluster}
\title{Estimate copy number}
\usage{
estimate_cnv_state_cluster(count, genome, copy_ratio, bkp, label, n_iter = 2)
}
\arguments{
\item{count}{the raw cell-by-bin read count matrix.}

\item{genome}{reference genome, "hg19" or "hg38".}

\item{copy_ratio}{the estimated copy ratio matrix from \code{normalize()}.}

\item{bkp}{the detected breakpoint positions from \code{calculate_CNV()}.}

\item{label}{a vector indicating normal cell and tumor subclone. Its length
should be the same as the row number of count matrix, and normal cell
should be labeled as "normal".}
}
\value{
Return a list of the following objects:
\itemize{
\item "CN_state": a cell-by-bin copy number state matrix, where elements
are one of 0.5 (copy number loss), 1 (copy number neural) and 1.5 (copy
number gain).
\item "copy_ratio_adjusted": a cell-by-bin copy ratio matrix.
}
}
\description{
Estimate copy number
}
