% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_input}
\alias{generate_input}
\title{Generate input matrix from fragment file for AtaCNV pipeline}
\usage{
generate_input(
  fragment_file,
  cell_barcodes,
  genome = "hg38",
  output_dir = "./"
)
}
\arguments{
\item{fragment_file}{path of fragment file.}

\item{cell_barcodes}{a string vector giving barcodes that represent cells in
scATAC-seq expriment.}

\item{genome}{reference genome, "hg19", "hg38" or "mm10".}

\item{output_dir}{output directory}
}
\value{
a cell-by-bin read count matrix
}
\description{
Generate input matrix from fragment file for AtaCNV pipeline
}
