% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize read count matrix and de-convolute coy ratio signals}
\usage{
normalize(
  count,
  genome = "hg38",
  mode = "normal cells",
  thre_bin = 0.8,
  thre_cell = 1,
  thre_mappability = 5e+05,
  logFT = F,
  dlm_dV = 0.3,
  dlm_dW = 0.01,
  count_paired = "none",
  normal_cells = "none",
  cell_cluster = "none",
  cluster_method = "kmeans",
  K = 5,
  thre_correlation = 0.95,
  use_max_cor_cluster = F,
  thre_ncell = 30,
  cutoff = 3,
  output_dir = "./",
  output_name = "norm_result.rds"
)
}
\arguments{
\item{count}{a cell-by-bin read count matrix. Column names should be formatted (see
example data).}

\item{genome}{reference genome, "hg19" or "hg38".}

\item{mode}{a string indicating how to obtain baseline counts of each bin.
One of:
\itemize{
\item "matched normal sample": baseline is obtained from inferred normal
cells using a matched normal sample. Must provide \code{count_paired}.
\item "normal cells": baseline is obtained from normal cells.
Must provide \code{normal_cells}.
\item "all cells": baseline is obtained from all cells. Can be used when no other
 information is available and tumor malignancy is low.
\item "none": baseline is obtained from inferred normal cells selected using gene numbers.
}}

\item{thre_bin, thre_cell}{real number thresholds between 0 and 1. In count matrix,
bins with proportion of zeros greater than \code{thre_bin} and cells with proportion of zeros
greater than \code{thre_cell} will be removed.}

\item{thre_mappability}{an integer between 0 and 1000000. Bins with less uniquely
mapped positions than the threshold will be removed.}

\item{logFT}{logical}

\item{dlm_dV, dlm_dW}{parameters of dynamic linear model. Variances of the
observation noise and the system noise, respectively.}

\item{count_paired}{a cell-by-bin read count matrix of paired normal sample.
Only used when \code{mode} is set to "matched normal sample".}

\item{normal_cells}{a vector giving normal cells. Should be a subset of rownames of [count].
Only used when \code{mode} is set to "normal cells".}

\item{cell_cluster}{a vector indicating cell cluster identity of each cell,
or "none" if there is no predefined cluster.
Only used when \code{mode} is set to "matched normal sample" or "none".}

\item{cluster_method}{clustering method, "kmeans" or "hc". Only used when
\code{cell_cluster} is set to "none".}

\item{K}{number of clusters. Only used when \code{cell_cluster} is set to "none".}

\item{thre_correlation}{a real number threshold between 0 and 1. When selecting
normal cells using matched normal sample, clusters that have higher similarity
than the threshold will be kept.
Only used when \code{mode} is set to "matched normal sample".}

\item{use_max_cor_cluster}{logical, indicating whether to select only cluster
with the highest similarity or all clusters with higher similarities than
\code{thre_correlation}.}

\item{thre_ncell}{an integer threshold. Clusters with less cell number than the
value will be neglected in normal cells selection.
Only used when \code{mode} is set to "matched normal sample" or "none".}

\item{cutoff}{a real number. Copy ratios larger than this value will be set to
it.}

\item{output_dir}{output directory}

\item{output_name}{output file name, should be ".rds"}
}
\value{
Return a list of the following objects:
\itemize{
\item "norm_count": a cell-by-bin normalized read count matrix.
\item "baseline": a vector. Read count baseline of normal cells.
\item "copy_ratio": a cell-by-bin copy ratio matrix without segmentation.
}
}
\description{
Normalize read count matrix and de-convolute coy ratio signals
}
\examples{

}
