% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot CNV heatmap using ComplexHeatmap}
\usage{
plot_heatmap(
  copy_ratio,
  del_chrXY = T,
  cell_cluster = "none",
  cluster_method = "kmeans",
  K = 5,
  cell_annotation = "none",
  save_hc = F,
  output_dir = "./",
  output_name = "copy_ratio.png"
)
}
\arguments{
\item{copy_ratio}{a cell-by-bin copy ratio matrix. Column names should be
formatted.}

\item{del_chrXY}{logical, whether to exclude chromosome X and Y?}

\item{cell_cluster}{a vector indicating cell cluster identity of each cell,
or "kmeans" if you want to use kmeans to assign cell cluster, or "none" if
there is no predefined cluster. When cluster identity is assigned,
[cluster_within_group()] will be used to decide row hierarchy of the heatmap.}

\item{K}{number of clusters.}

\item{cell_annotation}{a vector of cell annotations to be shown on the left
side of heat map, or "none".}

\item{save_hc}{logical, whether to save hierarchical clustering result?}

\item{output_dir}{output directory}

\item{output_name}{output image name, should be ".png" or ".pdf"}
}
\value{

}
\description{
Plot CNV heatmap using ComplexHeatmap
}
